frontend/dc-dashboard-app/backend/pyproject.toml
[project]
name = "talk-to-data-backend"
version = "0.1.0"
description = "FastAPI backend for Talk to Data NL-to-SQL conversion"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "baml-py>=0.76.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
